refactored import/export utils structure
