+++ frontend/src/views/positions/import-export.js
@@ -11,12 +11,10 @@ import { useSelector, useDispatch } from "react-redux";
 import { ExportActionButton } from "../../components/export-button";
 import { ImportActionButton } from "../../components/import-button";
 import { Alert } from "react-bootstrap";
-import {
-    normalizeImport,
-    prepareSpreadsheet,
-    dataToFile,
-} from "../../libs/importExportUtils";
-import { prepareMinimal } from "../../libs/exportUtils";
+import { normalizeImport } from "../../libs/import-export/normalizeImport";
+import { prepareSpreadsheet } from "../../libs/import-export/prepareSpreadsheet";
+import { dataToFile } from "../../libs/import-export/dataToFile";
+import { prepareMinimal } from "../../libs/import-export/prepareJson";
 import {
     PositionsList,
     PositionsDiffList,
